backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0

# OAuth
authlib==1.3.0
httpx==0.26.0

# AI/LLM
openai==1.10.0
anthropic==0.18.0

# Background tasks
celery==5.3.6
redis==5.0.1

# CRM API clients
requests==2.31.0
requests-oauthlib==1.3.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2024.1

# Social Media APIs
facebook-sdk==3.1.0
instagrapi==2.0.0

# Testing (dev)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
